requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
distro==1.8.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
openai==1.3.3
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.0
smmap==5.0.1
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
